requirements.txt
# Scrapy and Zyte API integration
scrapy>=2.11.0
scrapy-zyte-api>=0.32.0

# JavaScript rendering for Consejo spider (no Zyte)
scrapy-playwright>=0.0.33
playwright>=1.49.0

# Optional: for Scrapy Cloud CLI deploy
# shub

# Local R2 smoke test (test_r2.py)
boto3>=1.34.0
python-dotenv>=1.0.0

# Image download + WebP upload (radarlicencias/r2_image.py)
requests>=2.31.0
Pillow>=10.0.0
